More Map Bug Fixes (#19110)
* Modifies the Kataphract airlock a little. Help me god I can't fix the
docking problem.
* Hegemony Corvette: Fixes bum wire and pipes. Adds K'laxan body armor.
Moves some stuff around to be in their respective departments.
* Elyran Corvette: Adds more jetpacks, flagpatches, oxygen tanks, and
GPSes.
* Hailstorm ship: Adds fuel tanks, blast doors on the crew compartment,
very limited extra medical supplies, a toolbelt, and pistol holsters.
* Headmaster ship, Ranger gunboat: Fixes airless tiles in the Grauwolf
chamber and adds a button to the inside of the Ranger shuttle.
* Antag areas: Adds legbraces, neckbraces, jagmasks, and a couple
exostellars to the merc zone. Adds some storage (holsters, pouches,
duffels) to the ninja zone.
* Wrecked NT ship, overgrown mining asteroid: Adds areas to the
space-exposed parts of the asteroid to fix problems with gravity and
floating.
* Cosmetic changes to the overgrown mining asteroid and Hailstorm so
they look more lived-in.