Makes Wounds Appear on the Health Doll (#19111)
Adds wound overlays to the health doll. Basically ports [what I did on
Baystation12](https://github.com/Baystation12/Baystation12/pull/33765)
in a roundabout way.

**Why?**
Personally, it felt pretty jarring _not_ having them. More broadly, it's
a nice indicator of how hurt your character is and makes the health doll
a little less misleading about how beat up you are (as it only displays
the level of pain a certain body part is in).

**Did you test this?**
I did. It works fine with Humans, Skrell, Machines, Unathi, Diona, and
Tajara. If there's species I missed, it shouldn't be too tricky to go
back and check those, too.

**Some examples**
Let's say you've been stabbed with a kitchen knife a few times. That
might look a bit like this.

![image](https://github.com/Aurorastation/Aurora.3/assets/77447558/aa0cef03-aba9-407b-8002-cba8856b5290)
And once it's bandaged, it'd look like this.

![image](https://github.com/Aurorastation/Aurora.3/assets/77447558/60da7c3b-100e-48f1-ac6e-7d8cb168390b)

Co-authored-by: Nyello <karljohanssonpublic@gmail.com>