Moghresian Meat & Unathi Food Nutrition Improvement (#19135)
This PR creates a new subtype of meat, "moghresian meat", specific to
animals which are located in Uueoa-Esa.

It also adds "synthetic" blood to the biogenerator, for use in the many
recipes which require blood (I did this because it's actually impossible
to get blood from cows, chickens, etc - turns out you can't draw blood
frim simplemobs, and people have been either using their own blood, or
making monkey cubes and drawing blood from them, which is WEIRD.
Mechanically the "synthetic" blood is just generic blood. I have no idea
if vampires can abuse this, but it should be no different from them
making monkey cubes in the biogenerator and eating them, so... hopefully
that's ok?)

~I have also done a once-over on the reagent values of all the food in
the Unathi foods dm, and fixed the reagent value for them. Some foods
had far fewer reagents than their constituent ingredients, which
bothered me. I also fixed their flavours so you get nicer messages
instead of "some sort of protein"

I noticed that this food reagent issue is actually prevalent in ALL the
food .DMs. A good example of this is Black Eye Gumbo, which has no
reagents listed. It can be eaten in a single bite and tastes like
nothing, and cannot be ground. I think that some people have added foods
under the impression that the reagents of the ingredients are added by
the cooking process, when this doesn't actually seem to be the case. I
fixed this just for the Unathi foods, but I don't have the time to go
through them all.~

EDIT: Turns out the reagent thing was a red-herring caused by
adminspawning food. Fixed now!