Scarab Salvager, Izharshan Shuttle, Hivebot Hub tweaks & fixes (#19142)
This is a variety of fixes and changes for the Scarab Salvager and
Izharshan Shuttle offships, plus a number of tweaks to the Hivebot Hub
offsite.

**Scarab Salvager**

- Portside CO2 console works, now.
- Windows are replaced with proper variants to fit the surrounding
walls.
- Balance change, replaced the Xanu SMG in the armour with a pump
shotgun, so there's at least one longarm you can't put into a bag to
justify slinging over the armoured suit. Comes only with regular shells,
beanbags, and stun shells - but you can print slugs if you're so
inclined.

**Izharshan Shuttle**

- Fixed airlock. You can now cycle it, and it docks properly!
- Various small QoL changes. Removed shuttle floors, carpet should look
nicer.
- Power is shifted from RTGs to a mini-fusion that can actually keep up
with power consumption.
- Propulsion makes a little more sense now. It's still very slow as
ships go, but I suppose that's the trade-off for being able to dock.
- Only balance change is the addition of four seismic charges to the
ship. My thought is that, since a lot of ships lack docking ports and
this is an offship that specialises in boarding actions, it'd be much
more fashionable to blow a hole in the side of your quarry with
improvised explosives than it would be to go through the 15 step process
of disassembling a wall, or looking around for ages for an airlock.
Conveniently, when the charges reduce reinforced walls to a girder, an
energy cutlass can slice the girder apart.
**Note:** there's an odd bug in which telecommunications for this ship
doesn't work until you clear all of the filtering frequencies with a
multitool, but it _does_ work if you do that. If I find a fix I'll
commit it, but otherwise I'd like to get this through quickly so the
offship is at least mostly functional.

**Hivebot Hub**

- Adds drastically more hivebots. People keep expecting a horde of them
with this offsite, which leads to them overpreparing and finding
disappointingly little resistance after they do. This should give them a
more satisfying challenge. The beacon is still the weakened version.
- Moves the short-barrelled assault rifle to the 'end' of the map, since
it used to trivialise the hivebot threat.
- More than halves the amount of phoron in the offsite. 8 full canisters
was a little much.