Coalition Ship 'Purge' from Badlands (#19167)
The Badlands sector has alot of available awaysites/ships, but some
don't fit as well as others.

Coalition Gunship: Badlands far away from Coalition, while anti-pirate
activity could be argued as the reason, the Badlands has a variety of
military ships from other nations, that are seen far less often. For
this reason, this will not be set to spawn
Xanu Frigate: Same as above
Coalition Surveyor: The Badlands is indeed a place with unusual fauna,
but this should be handled by other survey vessels that could arrive,
the Coalition is focused on the 'northern' frontier far more than the
south.

Edit: Rangers and survey ship now in spawnpool but at reduced chance,
xanu frigate still removed

---------

Co-authored-by: Ben10083 <Ben10083@users.noreply.github.com>