Taj Jackboots Fix (#19361)
For the back sprite of the jackboots, one of the boots didn't align with
the overall taj sprite while the other did. For the toeless variant, the
back sprite didn't align with the taj sprite at all.

History will never know if I messed it up and only noticed it now or if
someone else did this.

Co-authored-by: WickedCybs <wickedcybs@gmail.com>