Bridge Crew & Zeng-Hu Corporate Reporters may now use Zeng-Hu augments (#19337)
Bridge Crew could already access some, but not all of the Zeng-Hu
exclusive augments. If this was intentional, I will roll back adding
them to the rest. Corporate Reporter, however, could formerly access
none - they may now access all, seeing as they are also a role that is
now employed by Zeng-Hu (but was not at the time of the augments being
added.)

---------

Signed-off-by: Pirouette <57296132+listerla@users.noreply.github.com>