Overcoming Verticality - Adds ramps, extends the medical lift and fixes stair gravity. (#19182)
Adds ramps, which function as dummy verticality like many of the stairs,
except they don't trigger the "Stairs are dangerous" response, allowing
certain areas to be more accessible.

Extends the medical lift to deck three, allowing a non-stair based way
to reach the upper department.

Fixes the "Stairs are dangerous" interaction not checking for gravity.

---------

Signed-off-by: Matt Atlas <mattiathebest2000@hotmail.it>
Co-authored-by: Matt Atlas <mattiathebest2000@hotmail.it>