Barista QoL Expansion (#19395)
Adds the ability to use pens to write names and orders on takeaway cups.

Allows you to click on syrup dispensers with a cup to dispense the
syrup, rather than having to pick them up.

Moves the chemical heater from the bar backroom to the front, so
bartenders don't need to walk away and come back every time they need to
make steamed milk.