Cult buff (#19386)
Added a single-use-per-round rune that can convert a mech into a cultist
mech, needs 2 minutes to do so and is single use, so max 1 cult mech per
round.
Added the ability to upgrade a pylon into turret mode using the cultist
blood, by clicking on the pylon with disarm intent.
Fixed do_after not using the callbacks at all (because it never passed
the parameter to do_after_detailed).
Some improvements to the pylons and runes code, DMDocs.

---------

Signed-off-by: Fluffy <65877598+FluffyGhoster@users.noreply.github.com>
Co-authored-by: Geeves <ggrobler447@gmail.com>