Return of the Paramedic (#19125)
This PR brings back the more accurate title for First Responders, of
Paramedic, to bring it in line with some snippets of the SOP and to be
more accurate to the role itself, as both Security Officers and the rare
and elusive Firefighting Atmospherics Technician can be considered first
responders.

also I just like it better.

Changes any mentions of first responder in gear/access/etc to Paramedic
or PARAMEDIC where applicable.