Long Live The Keiretsu: Adds the Zeng-Hu seniority mask (#19333)
![image](https://github.com/Aurorastation/Aurora.3/assets/59633984/9fa2017e-5ed4-4452-81ac-3cdd2997e328)

![image](https://github.com/Aurorastation/Aurora.3/assets/59633984/0226f988-ba7e-4bec-b8d4-9f1bf50e4d75)

![image](https://github.com/Aurorastation/Aurora.3/assets/59633984/6324c440-b345-441d-b4ae-bf474a2df2ba)

![image](https://github.com/Aurorastation/Aurora.3/assets/59633984/b30c74e4-8c10-450b-a1c4-06cfa406fa1f)

![image](https://github.com/Aurorastation/Aurora.3/assets/59633984/ee830ef5-d340-4fc7-b612-262129517f09)
As most who are aware of the veterancy plate know, it currently covers
the _entire_ front of the face, making it seem impractical (if not
physiologically incompatible) and the butt of many jokes. This PR adds a
less invasive alternative, makes both a variant of the integrated
medical HUD to make it more practical, and adds unique versions for
Vaurca Workers/Warriors, Bulwarks, and Skrell.