Tramp Freighter remap (#19353)
This is a full remap for the Tramp Freighter offship that intends to
bring the ship into closer alignment with newer mapping standards, while
playing the same role as the original map.

- Although the old Catspaw-class freighter is nostalgic, I don't think
the layout is really tenable now without changing it so much that it
might as well have been a full remap anyways. The tramp freighter plays
a very important role as the one universal offship - every whitelist,
every sector, any reason you can RP up for why you're doing anything
you're doing.
- The layout is very different, primarily to accommodate full docking
arms. The idea is that the freighter should always be able to move cargo
to and from other parties without having to traipse through space. With
this design, even if the other party doesn't have docking ports, they
can still just go over to the tramp freighter to conduct their business
in person. You can even just cater to people if you like, it has a
hydroponics bay and bartending equipment.
- All references to the ship that show up on IC mediums have been
renamed to 'Independent Freighter'.
- The cargo bay of the ship has been expanded. Four crates in the middle
of the room contain particularly valuable randomised items. The shuttle
is also designed to accommodate loading racks with merchandise like you
would on the merchant's shuttle.
- The guns in the armoury are half static, half random. The static
options are an energy rifle, energy pistol, and antique revolver - so
you'll always have *some* decent options.
- The ship generally is *full* of contraband, even outside of the
smuggling compartments. Don't get pulled over by a patrol ship, and if
you are, take a lot of care to clean up shop.
- The ship is deliberately weak in ship combat. It has relatively little
fuel and very punishing entry points - you can hit the bridge from any
direction but the aft, and doing so also cuts the gunnery. In my view,
the tramp should be the 'small fry' in ship combat, any pirate offship
should have a relatively easy time with it.