Gremorian Warriors/Workers Use Simple Melee AI (#19512)
Changes their smart melee var with the idea that warriors go for brute
strength over strategy and the workers are more simple in how they want
to spread the infestation. Makes the former a little different from most
of the usual greimorians and the latter less punishing when it comes to
spread.

---------

Co-authored-by: WickedCybs <wickedcybs@gmail.com>