Lore Description for Monkey Cubes and Boxes (#19521)
As titled. It sets the idea that the cubed animals aren't exactly the
same as their regular counterparts. This might be something for lore to
decide though, not me. So this PR should probably stay up for a bit.

For those too scared of clicking the "files changed" section. The
description is as follows.

"The manufacture of a cubed animal produces subjects that are similar
but have marked differences compared to their ordinary cousins. Higher
brain functions are all but destroyed and the life expectancy of the
cubed animal is greatly reduced, with most expiring only a few days
after introduction with water."

---------

Co-authored-by: WickedCybs <wickedcybs@gmail.com>