Allow Wireless Firing Pins (Disruptors) To Fire On Code Blue (#19434)
This adds code blue to the list of alert levels that allow wireless
firing pins to fire on lethal mode.

There's a cultural thing on Aurora where Code Red is never used, because
it stops the vote from being called. This means that during most high
intensity interactions, we stay on Code Blue, and the disruptor pistol
can't be used on lethal mode.

This simply adds code blue to the list of permissible codes for lethal.

---------

Signed-off-by: Crosarius <30341877+Crosarius@users.noreply.github.com>
Co-authored-by: Fluffy <65877598+FluffyGhoster@users.noreply.github.com>