Fixes SMES hacking (#19549)
~~In preparation for my grand political strategy to get SMES hacking
wires added to the aurora wiki, I made sure they were working.~~

It's on the wiki now so you have to merge this it's canon 🫡 

Changes:
* The RCON wire in a SMES now actually disconnects a SMES from RCON
(before it only prevented AI control).
* When the input wire is cut in a SMES, it will now correctly report
that the SMES is not charging on the UI.
* The Failsafe and Grounding wires of a SMES now correctly trigger the
red light when only one is cut, instead of both needing to be triggered.

---------

Signed-off-by: FlamingLily <80451102+FlamingLily@users.noreply.github.com>