Anomalist Alt Title For Xenoarchaeologist (#19559)
This re-adds the Anomalist alt-title for Xenoarchaeologist, alongside
some uniforms for them (They're just Xenoarch radios and uniforms which
are renamed to say "Anomalist")

The justification for this is that Xenoarchaeology and Anomaly research
are pretty distinct mechanically and lorewise. This isn't to say
necessarily that a Xenoarchaeologist is incapable of doing research on
anomalies or vise versa, and I don't want to tell people that they
can't!

However, I would like to have a way for Science characters with
specialisations in anomaly research to be able to distinguish themselves
from Xenoarchaeology, because their field wouldn't really have much to
do with doing anthropological research or fossils or archaeology,
really.