Welding Tool Surgery Fuel Usage Fix (#19597)
* Fixed welding tools using fuel in prosthetic repairs during the check
to see if it CAN do the surgery. Instead, the fuel when be consumed once
the surgery completes.

I _think_ this fixes the issue where welding fuel usage for prosthetic
surgery is bonkers, but I wasn't able to reproduce the bug on my local,
so keep an eye out for it still.