Expands the pirate base offsite to every sector (#19676)
Makes the pirate base offsite spawn in every sector, in the interest of
promoting more offsite variety. The cult base and freebooter ship are
both already universal, so there appears to be strong precedent. This is
also a good offsite for giving military offships something more to do
that doesn't involve the Horizon.