Hegemony Energy Rifle Bugfix (#19649)
Fixes Bug #19648

The Hegemony Energy Rifle has the E-Gun as a parent, instead of the
laser rifle, so it doesn't inherit all the single-fire-mode stuff that
the laser rifle has. I gave it all those attributes to make sure you
can't accidentally change the fire mode.

Edit: To clarify, this fixes the bug where you can change the fire mode
and it causes the gun to vanish.

---------

Signed-off-by: Crosarius <30341877+Crosarius@users.noreply.github.com>