Reduces explosion radius of standstill landmines (#19685)
This reduces the explosion radius of standstills. Currently, standstill
landmines blow a hole in the ship so large that it can take engineering
an extremely long time to fix it, even with full staffing. If a
standstill goes off inside service, for instance, nobody is using
service for a very large portion of the round, if ever - between the
venting, the broken floors, walling, windows, furniture, and machines. I
think it'd benefit as a tool without its users having to worry about the
insane blast radius disrupting more than they intended.

This reduces the heavy and light impact ranges, so they'll do less
structural damage, but maintains their devastation range - so, if you're
caught at the epicentre of the blast, you should still be about as
immediately critical as before. Lethality shouldn't be strongly
influenced, just structural damage. The blast radius is still large
enough that, in my testing, it's impossible to fully escape the blast.

Current blast radius:

![image](https://github.com/user-attachments/assets/0d3aa191-bf04-4fd8-98f6-a9e7cd08e836)

PR blast radius:

![image](https://github.com/user-attachments/assets/44499f07-f866-4626-9d51-c1fb604daf17)