The Tree Update: Makes it easier for Dionae to color their body markings, removes hairstyles and skin tone from them, makes the tail marking into an actual tail (#19589)
https://github.com/Aurorastation/Aurora.3/assets/59633984/76c7c7a4-fa06-457c-bf8c-0a9bda733f49

Basically, uses the species' list of skin color presets, if any, to add
a button for body markings to select the same presets. This is only
really useful for Dionae at the moment, or at least I imagine so. Skin
tones and hairstyles were also removed from Dionae, as figuring out how
to add selectable tone for body markings was too much of a hassle and
adding presets for hairstyles as well seemed redundant. I was thinking
of adding a "random body markings" proc to resolve the issue of newly
formed gestalts not spawning with any hair, but let me know if this is
desired.

I discussed this with the Diona lore maintainer beforehand, but the team
is free to change their mind or request more features at any time.