Honor, fire, burn thy fear: Adds a simple animal knockout technique to the Kis-Khan martial art (#19631)
Implements a disarm technique to Kis-Khan that "knocks out" simple
animals for a few seconds, allowing you to follow up with other attacks.
This is because shriekers are easy to encounter in the Untouched Lands
as a Kis-Khan master and are too difficult to get away from with Unathi
sprint, making it easy to get overwhelmed by multiple of them,
especially if you're in water. And dying to shriekers is, unfortunately,
not badass as a martial arts master.

Could also pave the way for other simple animal-based attacks for
Kis-Khan or other martial arts, if people are interested in that.