Remove the Crew Armory and the Leviathan (#19728)
end of an era


![image](https://github.com/user-attachments/assets/3a8c35b1-5dd6-4456-8f87-b807e61b8535)

![image](https://github.com/user-attachments/assets/723dd548-852d-48f8-9255-881aa7e6f71b)

![image](https://github.com/user-attachments/assets/157491eb-a1b8-465b-868a-d8df1c56aa41)

![image](https://github.com/user-attachments/assets/50c1ebbb-0d4f-40f0-bc4f-2675f126b8be)

![image](https://github.com/user-attachments/assets/e3e8c83f-4d77-4c81-aa55-263e90eb0c90)


----------------------------------------------------------------------------------

----------------------------------------------------------------------------------

----------------------------------------------------------------------------------

----------------------------------------------------------------------------------

changes:
  - rscadd: "Removes the Crew Armory."
  - rscadd: "Removes the Leviathan."
  - rscadd: "Updates mapmerge2 tool with most current tg version."



MATT'S EDIT: For whoever is reading, the Leviathan and the Crew Armoury
are being removed to try and lower the Horizon's passive militarization
a bit. We're going to try playing without the crew armoury, but in case
security stomps are unbearable and Ops can't properly fulfill the role
of "giving people guns", then ERTs will likely end up being readded.

There will be an IC explanation given for the Leviathan's removal. The
crew armoury one will probably be a straight retcon.

---------

Co-authored-by: DreamySkrell <>