Add a Callistean Lighter to the loadout options (#19765)
Take two; without panic closing a PR because it changed a file I didnt
want changed!

Adds a new item to the game and to the loadout options under 'zippos',
the Callistean lighter!

A cheap. Flimsy. Garbage lighter that Idris swears wasn't a money
laundering scheme.


![image](https://github.com/user-attachments/assets/cb2de131-d734-4633-9ae7-298e92528fa4)

This is intended as a warm up PR for me to re-familiarize myself with a
few tools; and to give something nice to the better moon.