Cryo heals below and to 60% BA instead of healing only at above 50% BA (#19790)
Allows cryogenics to heal the brain to 60% brain activity, at which
point natural healing brings it to where inaprovaline can take the
wheel.

This was intended initially by Kermit when she made the cryo PRs,
however by mistake the opposite was implemented, wherein cryo could only
heal when brain damage was above 50%. It was useless for it's original,
niche goal - allowing physicians to restore brain damage when no
surgeons or pharmacists are available.