Adds a small combustion turbine to Atmos maintenance. (#19767)
Repurposes a little bit of maintenance to give Atmospherics a singular
turbine to play around with. Lockers/crates/maint loot that already
existed in that room have been shifted to adjacent maintenance areas as
to not disturb the overall balance of goodies.

The rationale behind this is that it'll give Atmos Techs something to do
and to play with once thrusters are set up, while being completely
optional. Messing around with it is never the expectation, hence the dim
lighting, its tucked away nature, and how it totally lacks prior setup
beyond necessary construction. (The SMES is empty and not upgraded, no
canisters are in place and need to be manually found and filled
elsewhere, all the pumps and computers are unconfigured, etc.)

Not really intended to replace the SM or INDRA, as combustion requires
more maintenance and upkeep due to faster cooling, and actively burns
fuel to run - plus it's only a single turbine. But it does give bored or
curious Atmos Techs something else they can play with if they feel like
it, once all the round-start stuff is done.


![image](https://github.com/user-attachments/assets/51003fca-f291-4d9d-8b5c-06ca8c8800b5)



In conclusion:


![ezgif-1-bbc246f85f](https://github.com/user-attachments/assets/9be9735f-8431-4a2a-94c1-52badf8898b0)

---------

Signed-off-by: Joshie <41561534+furrycactus@users.noreply.github.com>
Co-authored-by: Fluffy <65877598+FluffyGhoster@users.noreply.github.com>