Mining QoL tweaks (#19741)
Remaps the mining refineries to have three separate refinery lines, to
match the three miner slots. It's also more compact, which should speed
up refining quite drastically.

This also adds an airlock variation of the SCC pressure tank and adds it
to the Spark. There's been issues with the Spark airlock needing to
perform an outward cycle to be able to cycle inward, so if it's
currently forced out it can jam going back in. With a pressure tank
rather than a canister, it can comfortably cycle inward twice without
any outward.