Replaces Supermatter Nitrogen with Helium (#19834)
Swaps out the four Nitrogen canisters in Engineering Hard Storage for
four Helium canisters instead.

Nitrogen is rarely if ever used for anything in the Supermatter except
emergency coolant (which there are already pre-cooled canisters for
inside the engine room) due to it having one of the lowest heat capacity
values in the game (20) - shared with Oxygen and surpassed by only Boron
(11) and Chlorine (5).

Helium is currently only relevant to the INDRA as a common fusion
product and reactant, but it does have a specific heat value of 80 - a
little lower than Hydrogen (100), but a lot higher than CO2 (30) which
are both already supplied in Hard Storage for use in the SM. This allows
for a little more potential diversity with setups, and gives Helium an
extra use.