Makes biogenerators visible again (#19883)
Makes biogenerators visible.

The problem was that the update_icon proc for biogenerators relied on
the initial icon_state being 'biogen' or 'biogen_small' to work, which
is an icon that didn't exist, so it showed up ingame through initialise
but not in an editor. A later PR changed that to make it show up in the
editor, causing it to break ingame. This renames the empty biogen
sprites simply to biogen and biogen_small to resolve the issue, so it
now shows up both ingame and in strongdmm.

The biogenerator in hydroponics and xenobotany on the Horizon had a
manually changed icon_state to get around this, which this also removes.

Resolves https://github.com/Aurorastation/Aurora.3/issues/19842