Fixes a few gun inhands (#19930)
Fixes the wield sprites of a few guns so they appear ingame. The
item_state definition seemed to cause the odd behaviour with the
Dominian rifles, and the Xanan rifles had a few mistyped icon names.

There seems to also be an issue in which inhand sprites for guns without
magazines don't appear, instead using the versions with magazines, which
is probably why the Dominian guns messed with the item_state in the
first place. Not tackling that here.