icon_state QoL improvements (#19915)
A few items now show up with appropriate sprites in StrongDMM. Achieved
by adding an icon_state if one was lacking, or by adding a generic icon
variant in the case of the energy weapons for the icon_state to
reference.


![image](https://github.com/user-attachments/assets/d2d7f871-adcd-4f5f-828a-2f1247da8858)

Added a crude sprite for the wasteland goggles whenever toggled off the
eyes, so it doesn't disappear when a player does so.