Adds an autocannon ship gun (#19927)
This adds the Goshawk heavy autocannon to the game, intended as a
generic gun emplacement available to be mapped into any appropriate
offship.

The AP shots are intended to be the weakest, but they do a lot of damage
on the off-chance they do hit someone. AP fires twelve shots, frag fires
nine, and HE only fires six - obviously ICly every salvo is the same
size, this is just for balance. As is currently, I'd say this is a
little stronger and generally a bit more versatile than the grauwolf.

**This is up for review.**

Credit to Noble for the spriting!