Life Could Be A Dream: Adds Vaurca Attendant as a Warrior subspecies (#19529)
![image](https://github.com/user-attachments/assets/0ee377f9-8321-4928-93d7-9b6d9a62fb83)
Basically, Attendants are now a subspecies of Warriors instead of a
subspecies of Workers, meaning they carry most of the traits over from
Warriors aside from the ones that have been modified by me. They also
have a unique appearance that should still work with Vaurca clothing,
instead of needing adjustments a la Bulwark. Hopefully, an acceptable
alternative to the Attendant subspecies that was previously offered in
#17550.

@desvenlafaxine